Add first working scloud module
This module is a modified version of the older sccrypt SClund module. The
modifications in the new module are:
- chang handling of key computation to support more then one key suite
- use HKDF as the key and locator derivation function
- remove the fine grained callbacks
- implement some buffer handling to compensate the removal of the callback,
  in particular during decryption
- add a function that reqturs the required buffer size during encryption
- fix some errors that causes wrong crypto block size calculation

The JSON module now uses cJSON instead of a complex parser. It's smaller and
easier to use.

Other modifications are on the TODO list: add AES-256, Twofish as key suites,
add the JNI modules