Add automatic pre-key management
The local conversation stores a counter of available pre-keys. The
registration initializes this counter to the number of registered pre-keys.
Each time the client receives a message type 2 indication a new Axoltol
pre-key setup the message handling functions decrements the counter.

The client checks this counter on each send and if it detects that the counter
is below a threshold then the send functions triggers creation and uploading
of new pre-keys and updates the counter.