Use context parameter if not null to re-hash the locator data
If the application sets the context parameters then the code uses the
data as 'salt' to re-hash the locator data. This avoid duplicate locator
data in case the application uploads the same file several times.
The application should fill the context data with some random data on
a per file basis.