Add and use specific error code in case of a race condition
If two users send message to each other at nearly the same time the client
may see that the Axolotl conversation already exists after the provisioning
server returned device info and the pre-key for the other user.

In this case return the new specific error code and the client tries to resend
the message using the existing Axolotl conversation.