Implement Id hash checks
The message sender hashes its long term identity key and the receiver's long
term identity key and adds the lowe 16 bit of both hashes to the message
enevelope. The receiver checks if the truncated hashes are available. If
this is the case the receiver computes own hashes and compares this with the
received data anmd reports mismatches.

This help to detect if one of the partners re-installed the application and
thus changed the long term identity key.