Prepare notify handling for changed reporting, change error report data
Prepare the SIP notify callback to handle a newer implementation of devices
available. This has to be agreed with the SIP server.

The error reporting on decryp now returns a hex string of the Axolotl
message. This provides information about the message type, length etc. It
does no reveal plain data, this is still encrypted.