Change handling of database error code handling to support multi-thread
The database classes are singletons and previouly stored the SQLite return
code in a instance variable. This is of coure not thread safe :-) . I changed
this so that calls of the DB functions can provide a address to an own
sql-result variable. If this variable is on the stack of the caller then
it's thread safe.

The modification was done in a way that functions that return the sql code
always return a thread safe code. The functions store the thread safe code
in the existing instance variable and return it to the caller.

The error code in the instance variable together with the error message
buffer is OK for testing an single thread use cases.