Add a map to store already seen device id notifications for a name
The code stores the device id content from a SIP NOTIFY using the full name
as the key into the map. If the SIP server sends the same string of device
ids again the code skip further processing.

The code only check for string equality, if the string differs the code
remove the map entry and add the new entry and then processes each device
id as usual.