Prepare to use alias and uid data for messaging
Add two functions to get additional information from SIP stack. The SIP
staack can provide a user's primary alias and the user's uid. During
migration the Axolotl lib uses the primary alias to lookup the conversation.

At a later time we may expand this to handle the user uid to peform lookups
for other aliase names, this however is only necessary if the user may change
the assignment of the primary alias, i.e. which of a user's alias name
is the primary alias.