Implement Emscripten support
This adds support for building with emscripten C++->JS compiler.

Files under interfaceApp/emscripten contain the implementation of
functions exposed to JavaScript. They are the equivalent of
interfaceApp/java for the Java client. The bindings exposed to JS
are in the `EMSCRIPTEN_BINDINGS` section of JSZina.cpp.

To work around the use of threads for processing the queue of sent
messages we process any messages in the queue immediately after they
are enqueued when compiling with the EMSCRIPTEN define. For this
reason it is recommended that the Zina routines are run from a
worker thread in JS to prevent blocking the browser main thread.

Zina expects to be able to call a function that makes synchronous
HTTP requests. This is implemented in `utils.js` in
interfaceApp/emscripten as the JS function `httpRequest`.  It in
turn calls a `syncRequest` JS function on the emscripten `Module`
object. This function must be implemented by the caller of an
emscripten compiled libzina to do a synchronous HTTP request. Note
that XMLHTTPRequest is asynchronous. Clients should use the npm
module `sync-request` from within a worker thread to prevent
blocking the browser UI.

Zina requires callback functions to be registered to handle
notification of messages sends, status, etc. From the JavaScript
side they are implemented as JS functions converted to pointers. eg:

    axo = new Module.JSZina();
    var index = Module.Runtime.addFunction(send_axo_message);
    axo.registerSendCallback(index);

Strings passed from JS to C++ are in UTF16. They need to be
converted to UTF8 and back which the calls `toUTF8` and `toUTF16` do
in JSZina.

Database files that libzina used are stored in a virtual filesystem
managed by emscripten. They are stored in the `/axolotl` path on
that virtual file system (see `mountFilesystem` in utils.js). The
caller of libzina can define the virtual filesystem as in memory or
indexed DB in a browser or as actual files in the filesystem in Node
JS.

Before calling JSZina.doInit the caller must ensure that the data in
index DB (if that is used for the virtual file system) are
synchronised with internal cached data by emscripten using
FS.syncfs. This must also be done when needing to commit data from
memory to the actual backing store after writes.

Random number entropy is made via changes to ZRTPCPP to call
JavaScript Web Crypto routines. In the case of browser based clients
where libzina is run on a worker thread we can't call Web
Crypto. Clients may, for example, create an object containing
functions that obtain entropy from a static array. The worker thread
would manage the data in this array and when it gets low post a
message to the main thread asking for more data and refill it from
the reply. The main thread would use Web Crypto to get this data.