Optimize the code to avoid network round trips on not known aliases or UUID
For ourgoing calls to PSTN (OCA calls) the server always returns a 404 Not found if we lookup the number. The server only
knows SC own numbers and names. To avoid many server requests store a special user info data structure for not found
names. If the use dials the same number again the name lookup has a cache hit. The lookup checks for the special user info
data and returns a not found in this case as well.