Add feature to amend the lookup-uri string to existing user info data using the addAliasToUuid function
Enhance and fix the inline documentation.

Add annotations to mark worker-thread functions, i.e. functions that must not run on UI threads in Android and maybe other
environment. To simplify the Java compilation I disable the annotations with a special '//**ANN** ' prefix. After copying the
file to the IDE project simply remove the special comments to enable the annotation/lint checks.