Remove the internal pre-key checks and management
Remove the client state to check number of available pre-keys and adding new pre-keys is necessary. Pre-key handling is
now part of the application. If another application get and thus removes one of our pre-keys then the server triggers a SIP
NOTIFY to the client to check the account data and status. The client reads the v1/me data, gets the new number of available
pre-keys, checks it and adds new pre-keys if necessary.