Release libzina as free and open source
ZINA is a secure messaging protocol for mobile devices.

This library, libzina, is an implementation of the ZINA protocol, and
is used in the Silent Circle suite of applications.

Silent Circle is releasing libzina under the Apache License Version
2.0 to help build a community around secure mobile messaging and to
encourage community review of the protocol and this implementation.

This implementation is by Werner Dittmann.

The design of ZINA is a collaboration between Werner Dittmann, Phil
Zimmermann, and Travis Cross.

Thanks-to: Ronald Krueger <rkrueger@silentcircle.com>

...for assisting in the release of this open source version.

Signed-off-by: Matt Neiderman <mneiderman@silentcircle.com>

...the General Counsel (CLO) for Silent Circle who authorized the
release of this library under the ASLv2 on 22 December, 2015.

History:

- An earlier version of this library was released under a different
  name on 2016-02-04 18:34:55 UTC and without the full project
  history.