Implement a conversation re-sync feature
If the ratchet cannot decrypt a message it reports this to the application. The application may now take actions,
for example to inform the user. As an additional option the application may now call the new reSyncConversation
function. This function resets the conersation crypto data and other states, creates a 'Ping' message and sends
it to the partner's device to force a re-sync of the ratchet data.