Implement a message caputure feature
This feature contains functions to store and retrieve message trace data. The feature stores a subset of message data to
provide information about the state of a message, for example if a receiver of a message sent a  delivery report or if the
client sent a read-receipt for a message.

The feature does not store any message content or sensitive data such as location information.