Extend group and member data structures with attribute and last modified fields
We will use these fields to have a better control of group and member attribute like Active, inactive etc and when we last
changed the data structure. The group and memger data management functions access and use this data.

Move 'Constants.h' up in the directory hierarchy because it does not only belong to the ratchet only.