NGA-580, NGI-536 Fix handling (commiting) of staged message keys
The staged message keys were not commited correctly therfore the decrypting failed for out-of-sync messages.
The functions now use a list to store the staged message keys, use this list during decrypting to get message
keys. The function commits the list of remaining message keys to the database after decrypting a message and
preparing possible other message keys.