Implement queued message sending to avoid losing messages
The SIP stack provides a limited number of SIP packet send slots. In case the client needs
to send a large number of messges, for example to a group, then it may happen that the
client runs out of SIP send slots. This happens if the SIP sever needs a long time to respond
to SIP packtes.

The new send function queues the necessary data to send a message in a list and empties
the list in a separate thread an takes care to send messages only if slots are available. The
new functions also requires a modified sendAxoData in the silentphone library.