Avoid attacks on convergent encryption
Convergent encryption (encrypting data using a symmetric key derived
only from the plaintext) suffers from confirmation attacks and
plaintext recovery when the encrypted data is a low-entropy variation
of a known plaintext.  When used without an authenticity mechanism, it
can also enable substitution of malicious attacker-controlled data in
place of a known plaintext.

The design therefore called for the use of a high-entropy salt value
not disclosed to the attacker to be involved in the generation of the
symmetric key for data encryption.  This was the motivation behind the
`contextStr` initialization parameter.  Products which use libzina
populate this parameter with secret high-entropy data.

Prior to this commit, however, the code here used this salt only in
the generation of the `locator`, the file name used for storing the
encrypted data.  This was an error.

Instead we want both the symmetric key and the locator to be derived
from both the plaintext data and secret salt.

In this code, we use HKDF [1] as a PRF to produce the symmetric key,
the initialization vector, and the locator.  Previously we were not
providing HKDF with a salt value.  In the absence of a salt, HKDF uses
a string of zeros equal in length to the output of the underlying hash
function.

With this commit, we feed into HKDF the high-entropy `contextStr` as
the salt.  HKDF uses this salt as the key input to the HMAC function
in the entropy extraction step.  RFC5869 supports our use of a secret
salt value, section 3.1:

    It is worth noting that, while not the typical case, some
    applications may even have a secret salt value available for use;
    in such a case, HKDF provides an even stronger security guarantee.

If the salt value is chosen as a per-use random nonce, as is currently
done in all products using libzina that we know of, then this is
believed to effectively negate all benefits and drawbacks of
convergent encryption.

Thanks to Tanja Lange and Sebastian Verschoor for pointing out the
error here.

Reported-by: Tanja Lange <tanja@hyperelliptic.org>
Reported-by: Sebastian Verschoor <srverschoor@uwaterloo.ca>

[1] https://tools.ietf.org/html/rfc5869