Change code to send delivery receipt, handle delivrey receipt in retry function also
Send delivery receipts only for real messages, not for command messages, otherwise we go into an endless loop
sending delivery receipts :-) .

Add a boolean ro the prepMessage* function to indicate if the message is a normal message or a command message.
The default setting is 'normal message'. Add a JSON key to the message descriptor when receiving a message to
indicate the message type. This is backward compatible.