Fix an issue with the own device name in the ratchet state data
The re-scan devices function now correctly sets the device name of its own device in its ratchet state data.
The getOwnIdentityKey can now use this information.

An application may get getOwnIdentityKey and check if the device name element is empty. If it is then it should
run the re-scan devices function (not in UI thread) and then use the information.