Unlock name map cache while running a network action
To avoid long locks on the name map cache we now release the lock when starting the network action and
get it again afterwards. Because of possible race conditions the handling of insert failures was changed as
well. Actually, insert returns fals only if the entry already existis in the map, thus this is not really an error in
our case. Just ignore this and return then requested data.