Don't store plaintext in the plaintext message receive queue if data retention checks fail
Fix the sequenece of actions to check for data retention, store plaintext in the plaintext message receive queue, and
send the delivery receipt. The new sequence now stores the plaintext data only if the data retention check returns
true, thus the client can proceed to process plaintext data.

The former sequence stored the data first, then checked for data retention, and removed the stored plaintext it if the
data retention check returned false. This opened a time window where the plaintext was stored in the receive queue
but not yet checked for data retention. If the application crashed or was forcefully terminated by the OS during this
time window then the retry/recover function took the stored plaintext and processed it, thus nullifying the data
retention checks.