Add dataRetention parameter to doIniti() to close a small time window of no correctly initialized local data retention flags
The caller of Java doInit() must provide the data retention flags in the same format as for the setDataRetention flags.
The native doInit() implementation then sets the data retention flags before it sets the transport (network) callbacks.
This avoids a small time window we the client could receive and process messages with not yet correctly initialized
data retention flags.