Remove internal R* JSON keys, add a simple dr-status indicator
The message attribute does not contain the R* flags anymore, these are internal elements. Replace them with
one dr-status indicator which is a bit mask (an integer) that contains the same information, however simpler
to test in the UI part. Instead of testing four flags it's now sufficient to just check one status field. If this status
field is zero then no R* flag is true.