Disable data retention by stubbing out data retention functionality
If SC_ENABLE_DR is not defined then data retention functionality
is disabled and code that performs the retention of data is not
included in the biuld.

Defining SC_ENABLE_DR enables data retention and includes the
code in the build.

SC_ENABLE_DR is undefined in this commit so data retention is
disabled.

(cherry picked from commit ca56e98)