Add a ratchet context id and implement ratchet wire format version 2
Introduce a ratchet context id to avoid error messages in case of ratchet re-keying. During
set-up of a new ratchet Alice creates a ratchet context id, stores it in Bob's ratchet context
and sends it together with the ratchet set-up data to Bob. Bob stores the context id in
Alice's context. Thus creation and storing the context id is part of the context set-up.

Each message contains this context id and each party checks it to verify that the message was
created with the current ratchet context. If the context ids don't match then it's an old
message and ZINA reports an error.

Because each message now contains the context id it was necessary to change the internal wire
format of the message. The new wire format uses proto-buffers only, not an internal binary
format anymore. Using proto-buffers gives us more flexibility for future enhancements.

The modifications are backward compatible, however older client which use version 1 od the
wire format cannot use or support the new context id.

During the changes for the new wire format also some cleanup to remove unnecessary shared
pointers and a few changes of internal APIs.