Add detection, reportingn and handling if identity key of a partner changes
ZINA detects and reports if a partner's client changed the long term identity key.  To report this every message deycriptor
that ZINA add to the UI callback has a new JSON field 'idkc' which is true if the id key was changed, false if not. ZINA also has
a new API that the UI code can use to set the id key to verfified state if the UI offers the user to verify the key manually. ZINA
also checks and verifies the id key if the users call and ZRTP can verify the keys. The long term identity key of a client changes
on a re-install of SP, not on a normal update. Of course ZINA also handles a new/first installation of SP as a id key change.

On a normal re-key operation the loing term identity key usually does not change. To add more security here as well ZINA now
has an enhanced check of the pre-key material that both parties use to re-key the ratchet. To perform a normal re-key the UI
code must now user the command 'reKeyAllDevices' and not 'removeAxoConversation' as before.

The command 'removeAxoConversation' only remove all ZINA converstaions (ratchet data) of the user but does no perform a
rescan of the user's devices anymore. The 'reKeyAllDevices' command is more lightwight and performs the id key checks.

Only if the UI code likes to reset all ratchet data and then re-create the conversation then the UI code should use the command
'removeAxoConversation'  followed by a 'rescanUserDevices' command. This sequence forces ZINA to handle the user as a
"new conversation" and thus sets the 'idkc' fields to true.