Fix issue when preparing change set for a device - use group id instead of group receipient (member)
Group change set are meant for the whole group, usually not just for a single member.
The exception is a specific handling if the client receives an unexpected group command for a group it does
not know. In this case, and in case of ACK response we send a change set to one specific device only.