Fix handling of pre-hash check
Compute and check the pre-key hash only if we created a new ratchet context. If the client got multiple
type 2 (setup) messages then don't compute and check the message hash for the following setup
messages. After the first message the pre-key data is not longer valid and not in use anymore.