Remove calls to SQLiteStoreConv::getStore and use the store pointer in AppInterfaceImpl, cleanup message key handling
In a first step remove calls to SQLiteStoreConv::getStore for the normal functions, not for trace functions yet. This change helps
to change the implementation of the data if this becomes necessary or to support multiple application interfaces with different
database. It's also much cleaner to use a common pointer to a database instance.

The cleanup of message key handling ZINA does not store any message keys while decrypting a message, only after a successful
decryption, containes in the main transaction where ZINA stores the context. Before the cleanup ZINA store message keys during
decryption in a specific case - this is now gone.