Change behaviour functions to re-scan and re-key devices
Before the changes to queue received and messages to send this two function where blocking, synchronous functions, thus
they returned to the caller after rescanning/re-keying was done. This changed with the changes to introduce queuing and
the functions became non-blocking. This new behaviour confuses the caller and lead to other issues.

The fixes in this commit re-establish the blocking, synchronous behaviour.