Add new versions of prepare message functions, depracte old versions
The new versions of the prepare message functions use unique pointers instead of shared pointers. Unique pointer use
less resoures and also show the ownership of the objects. The name of the new functions are:

- prepareMessageNormal(...)
- prepareMessageSiblings(...)

and replace the functions prepareMessage(...) and prepareMessageToSiblings(...) respectively. The parameters of the
new versions are the same as for the old versions.