Fix copy/paste error that breaks old API
When `AppInterfaceImpl::prepareMessage` and
`AppInterfaceImpl::prepareMessageToSiblings` were deprecated, a change
was made to `prepareMessage` in the call to `prepareMessageInternal`
that caused the `toSibling` parameter to be set to `true` rather than
`false`.  This breaks messaging.

This appears to be a copy/paste error from a similar change made to
`prepareMessageToSiblings`.

This was introduced in commit 84c692d52b02ff71c504d9268d6515a6a0eb24de.

Ed noticed this after we bisected down to this commit.

Thanks-to: Ed Fine <efine@silentcircle.com>
Thanks-to: Eric Turner <eturner@silentcircle.com>