Synchronize a group member's new additional device
The synchronization works as defined in the group chat protocol:

<quote>
As additional errata, the proposal should also specify the handling
for what happens when a new device is noticed for a group member.
Alice is sending a group message to Bob and notices that Bob has a
not-previously-seen device, BobD3.  Alice needs to notice this and
send BobD3 the full set of group onboarding metadata including her own
current vector clocks for each piece of metadata.

Note that other group member devices will also do this.  So BobD3 will
receive the full set of group metadata from every device in the group
that participates further in the conversation.

This is a bit brute-force, but does ensure that BobD3 ends up in a
consistent state.  We can't just have BobD3 broadcast an ACK to the
whole group because other group members wouldn't know whether Bob had
received an out-of-date or conflicting state from the member that sent
him the metadata first.
</quote>