Change error handling when sending groug messages
The new error handling skips over a problematic device, i.e. if a member
has no device, the device has no pre-keys etc, and continues to send the
message to the other devices of the member (if available) and also
continues to send the message to other members. However, the group send
remembers the error code and reports it to the caller. Thus if
|applyGroupChangeSet| or |sendGroupMessage| return an error code then it
detected /at least one/ problem when sending a message to the group members.