Add functions to manually burn a group message
The UI code should call `burnGroupMessage(const string& groupId, const string& messageId)`
to manually burn a group message. As discussed only the author of a group message sholud
be able to manually burn the message. ZINA does not (cannot) check if the user really is
the author of the message, thus the UI code needs to check this.

After calling the `burnGroupMessage` the UI code should either call `applyGroupChangeSet`
or just send a group message to inform the other members.

ZINA constructs a JSON string which contains the following keys:
- "grp":    "rmsg"    # GROUP_COMMAND, REMOVE_MSG
- "grpId":  "..."     # GROUP_ID
- "msgdId": "..."     # MSG_ID
- "mbrId":  "..."     # MEMBER_ID
- "cmd_time": "..."   # time stamp

The UI may use the member ID to show who burned the message.