Change handling to detect sibling devices
The provisioning server sends (via the SIP server) a SIP NOTIFY with an
'Event: x-sc-refresh-provisioning' attribute. The SIP stack sends this NOTIY
to the UI code to trigger the necessary updates. The provisioning server sends
this NOTIFY also if the user adds or removes a device the the account (sibling
device), thereforethe UI code _must_ start a device rescan for the own account
to learn the new sibling devices (or remove old devices).

The new sibling detection handling relies on this rescan. If the UI does not
start the rescan ZINA cannot detect new sibling devices.