Add conflict resolution in case of conflicting ratchet setup messages
Check the message type and the conversation state to detect a conflicting ratchet setup:
- if the conversation has a valid A0 then this client created the conversation and assumed Alice role
- if we now receive a type 2 message (setup message) then our partner also created a conversation
  assuming Alice role - and this is not correct

In this case we have to perform some conflict resolution and manage the ratchet contexts
If our local long term identity key (public part) is greater than the partner's key then we won :-) and
use our ratchet data. Otherwise we lost and just use the partner's setup message to re-set (re-key)
the ratchet context.

However, to avoid losing messages the winner has to prepare a secondary ratchet context, initialize
it, save it, etc and use it if necessary.