Make pending group change sets persistent
The group update function now store pending group change sets in database as
well as in an internal cache. In case of a crash of if the user terminates
the client fills the internal cahe from the database after a restart.

Update and optimize handling of the WaitForAck database records.