Split sendSimpleMessage into two steps in emscripten backend
sendSimpleMessage becomes prepareSimpleMessage and the call
to doSendMessage is moved out into another function to be
called by the JS front end.

sendSimpleMessage takes an array of strings as an argument
and it is populated with the JSON string containing the
transport id's and device information for the messages
being sent to specific devices.

This allows the JS front end to store information on
the transport id and device so if it later gets an
error via the state callback it can associate the
given transport id with the actual message sent.

The JS front end then calls doSendMessage passing
it an array of strings containing the hex encoded
transport id's to send.

The JS front end only ever deals with hex encoded
transport id's. The libzina code here converts that
to and from uin64_t as required.