Fix required version of sqlcipher to v3.4
When `-DANDROID` is not provided, we require that `sqlcipher` be
present in the environment.  The check was requiring at least
sqlcipher v3.7, but sqlcipher's latest release is v3.4.2.  That is,
there was confusion between sqlcipher and sqlite versions.  This
commit corrects the version required, and sets it to require at least
sqlcipher v3.4.