Add support for building with clang sanitizers
Clang supports sanitizers which add runtime checks to detect program
misbehavior.  In this commit we add support for building with:

- `UBSAN`: UndefinedBehaviorSanitizer detects the use of certain kinds
  of undefined behavior such as overflow when casting to or from
  floating-point types, out of bounds shifts, pointer arithmetic
  overflow, etc.

- `ASAN`: AddressSanitizer detects certain kinds of memory access
  errors such as double-free, invalid free, use-after-free,
  out-of-bounds access, etc.

- `TSAN`: ThreadSanitizer detects data races between threads

- `SSSAN`: SafeStack detects stack buffer overflows

- `CFISAN`: Control Flow Integrity checks detect certain kinds of
  undefined behavior which may allow an attacker to affect control
  flow

To build with these sanitizers, set the appropriate defines when
running cmake, e.g.:

    CC=clang CXX=clang++ cmake \
      -DUBSAN=ON -DASAN=ON -DTSAN=OFF -DSSSAN=ON -DCFISAN=ON \
      -DCMAKE_BUILD_TYPE=Debug .

By default, all sanitizers are disabled.

All sanitizers are configured such that they will cause the
application to exit if a problem is detected.

Because these sanitizers require the Clang compiler, unless this is an
automatic build, we detect the compiler in use, and exit with an error
if Clang is not selected.  In interactive builds we also test the
version of the compiler and enable flags appropriate to each Clang
version.  Clang v3.8 and later are supported.

It's not possible to run both ASAN and TSAN in the same build, and an
error will be generated if this is attempted.

The UndefinedBehaviorSanitizer usually includes a runtime that
generates diagnostics when an error is found.  However it can be
configured to simply trap.  When UBSAN is defined, we include the
runtime in debug builds, and in release builds we do not,
automatically.

UndefinedBehaviorSanitizer, SafeStack, and CFI have low overhead and
are meant to be run on release builds.  AddressSanitizer and
ThreadSanitizer, however, have considerable overhead and may not be
appropriate for all environments.  We output a warning when these are
defined for release builds.

Note that CFI requires link-time optimization (LTO).  This is enabled
automatically when `CFISAN` is defined.

ref: https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html
ref: https://clang.llvm.org/docs/AddressSanitizer.html
ref: https://clang.llvm.org/docs/ThreadSanitizer.html
ref: https://clang.llvm.org/docs/SafeStack.html
ref: https://clang.llvm.org/docs/ControlFlowIntegrity.html