Define undefined `double`->`int` casts in cJSON
An undefined `double`->`int` cast present in cJSON affects our ZINA
`contextId` value.  This affects our determination of whether we share
a security association with the remote party, and affects
compatibility between platforms, as described below.

When cJSON is asked for a number from a JSON field as an integer, it
parses the number into a double precision floating point value, then
casts that `double` value to an `int`.  This cast is undefined in C if
the `double` value is greater than `INT_MAX` or less than `INT_MIN`.

There's no clear way that cJSON's caller could prevent this.  We're
calling cJSON to parse the document after all; we don't know yet how
big or small the numbers contained therein are [1].  Therefore it's
reasonable to expect cJSON to provide defined behavior here.

However, as we've been relying on this undefined behavior, we need to
consider the impact to deployed software.

On x86_64 with gcc (v7.3) `-O0`, the `double`->`int` cast returns
`INT_MIN` when the value is greater than `INT_MAX` or less than
`INT_MIN`.  However with `-O1` or greater, it clamps the return value
at `INT_MAX` when the value is greater than `INT_MAX`; values less
than `INT_MIN` still return `INT_MIN`.

Clang (v6) on x86_64 also returns `INT_MIN` for `-O0` when the value
is greater than `INT_MAX` or less than `INT_MIN`.  But with `-O1` or
greater, it returns zero in these cases.

On ARM, under release build configurations for Android and iOS, for
positive values greater than `INT_MAX`, the return value is believed
to be clamped at `INT_MAX`.

Under Emscripten when generating asm.js, positive values greater than
`INT_MAX` seem to return a negative integer corresponding to the two's
complement value of the least-significant four bytes.  For WebAssembly
code generation, see `BINARYEN_TRAP_MODE=clamp` in the Emscripten
documentation.

To fix this in cJSON with the least impact on deployed software, and
to achieve compatibility between platforms, we're going to adopt the
optimized GCC/ARM behavior of clamping the cast value at `INT_MAX` and
`INT_MIN`.

In libzina, this undefined behavior affects the conversation
`contextId`.  This value is used to determine whether two devices
share the same security association.  If they do not and we proceed
anyway, we will end up with undecryptable messages.

The `contextId` is an unsigned 4 byte value, and the most significant
2 bytes are generated uniformly at random.  This value is serialized
to JSON, then deserialized as a signed integer.  When the original
unsigned value is greater than `INT_MAX`, which it is half the time,
this cast was undefined.  The undefined value caused the comparison of
our `contextId` with that of our partner to be undefined.

When two communicating devices agreed about how to interpret the
undefined behavior, this would have caused the `contextId` to match
more often than it should have.

On the other hand, when the two devices interpreted the undefined
behavior in different ways, the parties would have instead believed
their security association to be invalid too often.

This commit takes the first step in addressing this.  We define the
previously undefined behavior so that it is consistent across all
platforms.  This continues to allow `contextId` values to match more
often than they should.

[1] The JSON specification allows for floating point numbers of
unlimited size.  Commonly-accepted implementations support double
precision numbers within the IEEE 754-2008 limits.

ref: https://www.json.org/
ref: http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf
ref: https://tools.ietf.org/html/rfc7159#section-6
ref: https://github.com/kripken/emscripten/wiki/WebAssembly
ref: https://github.com/kripken/emscripten/issues/4625