Validate deserialized `contextId`
If the deserialized `contextId` has been clamped at `INT_MAX`, then we
may incorrectly report a `contextId` match due to the aliasing.  As
this may lead to undecryptable messages, we'll check for this and log
a warning.

Because the JSON field representing the `contextId` is deserialized
originally as a `double`, the value could theoretically be negative.
This cast is well defined, but should never happen, so we'll add a
warning here as well.

The original code here did a bitwise AND of the deserialized
`contextId` value with `UINT_MAX`.  As the value is a signed integer,
this bitwise AND is a no-op.  Presumably the thinking was that this
deserialized value could be larger than `UINT_MAX`.  In this commit,
we've removed this operation.