Add utility to get unsigned integer JSON value
This adds a utility `getJsonUInt` that returns a numeric field value
from a JSON object as a `uint32_t`.  cJSON parses the value as a
`double`.  A double precision floating point value in IEEE 743-2008
precisely represents integers up to 2^53.  We test to ensure the value
is not more or less than can be represented in a `uint32_t`.  If it
is, then we clamp the value.