Avoid clamping of new `contextId` values
We were generating the `contextId` with the most significant two bytes
set to uniformly random values.  However, due to the previous use of
undefined behavior in cJSON when deserializing this value as a signed
integer, half of the possible values for this field would be clamped
to `INT_MAX`.  For compatibility with deployed software, we've since
made this clamping the defined behavior.

For new conversations between clients running new software, we want to
avoid this clamping as it results in aliasing of the `contextId`
value.

In this commit, we mask out the most significant bit of
newly-generated `contextId` values so that they will not alias due to
the clamping.  However, note that the random part of this field is now
only 15 bits wide, so much more occasional aliasing is still likely.