Extend context identifier by 32 bits
Our context identifier (`contextId`) only contains 15 bits of entropy.
Even aside from the now-resolved clamping issue, this means that we
would expect this value to alias more frequently than we would like.

To make these collisions less likely, we'll extend our context
identifier by 4 bytes by adding a `contextId2` field.  This entire
field is filled with uniformly random bits.  This brings our combined
context identifier to 47 bits of entropy.

When this field is supported by both sides, both it and the
`contextId` must match before we assume that we share a context with
the other side.